talhaaltinkaya-b1-logs (added)
I will also send you an e-mail about this subject. 

Although I have run Node for a month, my logs only appear to be 17-18 days old. I did not touch the machine in any way during this period. I just followed it on the dashboard. And before I stopped the node, I checked it on the dashboard. I saw the text "1 mon" under uptime there and stopped it. But I get the same records even though I have tried a few times getting log. To check this, you can check the page on my node in the dashboard below:

https://dashboard.testnet.concordium.com/node/939f128d24c636f1

I would be glad if you give me feedback on the subject.